Add user res api

+++ FinTrack.Services/FinanceService.cs
@@ -6,14 +6,16 @@ using FinTrack.Services.Dtos;
 using FinTrack.Services.Mappers.Contracts;
 using Microsoft.Extensions.Logging;
 using FinTrack.Services.Exceptions;
+using FinTrack.Services.Context;
+using FinTrack.Services.Context.Contracts;
 
 namespace FinTrack.Services
 {
     public class FinanceService : AbstractService, IFinanceService
     {
-        public FinanceService(ILogger logger, IMapperFactory mapperFactory
-            , IDataContextManager dataContextManager, ContextLocator contextLocator)
-            : base(logger, mapperFactory, dataContextManager, contextLocator)
+        public FinanceService(ILogger<FinanceService> logger, IMapperFactory mapperFactory
+            , IDataContextManager dataContextManager, LocalizationContextLocator localization, IContextLocator contextLocator)
+            : base(logger, mapperFactory, dataContextManager, localization, contextLocator)
         {
         }
 
@@ -25,19 +27,29 @@ namespace FinTrack.Services
             var mapper = MapperFactory.GetMapper<IFinanceMapper>();
 
             var finance = mapper.MapFromDto(currencyDto);
-            financeRepository.Add(finance);
+            finance.CreatedDate = DateTime.UtcNow;
+            finance.UpdatedDate = DateTime.UtcNow;
+            await financeRepository.Add(finance);
 
             Logger.LogInformation($"FinanceService.AddFinance completed");
         }
 
-        public async Task Delete(int id)
+        public async Task Delete(Guid id)
         {
             Logger.LogInformation($"FinanceService.Delete started");
 
-            var resourceProvider = ContextLocator.GetContext<LocaleContext>().ResourceProvider;
+            var resourceProvider = LocalizationContext.GetContext<LocaleContext>().ResourceProvider;
             var financeRepository = DataContextManager.CreateRepository<IFinanceRepository>();
 
-            var finance = financeRepository.Delete(id);
+            var finance = await financeRepository.GetFinanceById(id);
+
+            if (UserContext.Id != finance.UserId)
+            {
+                Logger.LogWarning($"FinanceService.Delete there is no access to the data.");
+                throw new ValidationException("No access data.", resourceProvider.Get("NoAccessData"));
+            }
+
+            await financeRepository.Delete(id);
 
             if (finance == null)
             {
@@ -50,15 +62,21 @@ namespace FinTrack.Services
             Logger.LogInformation($"FinanceService.Delete({id}) completed");
         }
 
-        public async Task<FinanceDto> GetFinanceById(int id)
+        public async Task<FinanceDto> GetFinanceById(Guid id)
         {
             Logger.LogInformation($"FinanceService.GetFinanceById({id}) started");
 
-            var resourceProvider = ContextLocator.GetContext<LocaleContext>().ResourceProvider;
+            var resourceProvider = LocalizationContext.GetContext<LocaleContext>().ResourceProvider;
 
             var financeRepository = DataContextManager.CreateRepository<IFinanceRepository>();
 
-            var finance = await financeRepository.GetFinanceById(id);  
+            var finance = await financeRepository.GetFinanceById(id);
+
+            if (UserContext.Id != finance.UserId)
+            {
+                Logger.LogWarning($"FinanceService.GetFinanceById there is no access to the data.");
+                throw new ValidationException("No access data.", resourceProvider.Get("NoAccessData"));
+            }
 
             if (finance == null)
             {
@@ -92,11 +110,17 @@ namespace FinTrack.Services
         {
             Logger.LogInformation($"FinanceService.Update started");
 
-            var resourceProvider = ContextLocator.GetContext<LocaleContext>().ResourceProvider;
+            var resourceProvider = LocalizationContext.GetContext<LocaleContext>().ResourceProvider;
 
             var financeRepository = DataContextManager.CreateRepository<IFinanceRepository>();
 
             var finance = await financeRepository.GetFinanceById(financeDto.Id);
+
+            if (UserContext.Id != finance.UserId)
+            {
+                Logger.LogWarning($"FinanceService.Update there is no access to the data.");
+                throw new ValidationException("No access data.", resourceProvider.Get("NoAccessData"));
+            }
 
             if (finance == null)
             {
@@ -104,9 +128,11 @@ namespace FinTrack.Services
                 throw new ValidationException("Finance was not found.", resourceProvider.Get("FinanceWasNotFound"));
             }
 
+
             var mapper = MapperFactory.GetMapper<IFinanceMapper>();
 
             mapper.MapFromDto(financeDto, destination: finance);
+            finance.UpdatedDate = DateTime.UtcNow;
             await DataContextManager.SaveAsync();
 
             Logger.LogInformation($"FinanceService.Update completed");
